+++ templates/layout.php
@@ -1,5 +1,12 @@
 <div id="erp-wrapper">
     <h1 id="erp-header"><span ng-bind="pageTitle"></span></h1>
+    <div class="erp-tool-bar text-right">
+        <button type="button"
+            class="btn btn-default btn-sm ng-cloak btn-refresh-page"
+            ng-disabled="isBusy"
+            ng-click="refreshPage()">
+            <span class="glyphicon glyphicon-refresh"></span></button>
+    </div>
     <div id="erp-content" ng-class="{busy: isBusy}" ng-cloak>
         <div class="loading-overlay" ng-if="isBusy">
             <div style="position: relative; display: inline-block">